+++ packages/arc-utils/src/index.ts
@@ -2,3 +2,4 @@ export * from "./build-arc-file";
 export * from "./copy-shared";
 export * from "./set-env-vars-circleci";
 export * from "./build-preferences-file";
+export * from "./rename-index-files";